+++ runtime/src/main/3.3.2-scala/org/apache/spark/sql/ShimUtils.scala
@@ -77,6 +77,18 @@ object ShimUtils {
   /**
    * Registers functions with spark, Introduced in 0.4 - 3.2.0 support due to extra source parameter - "built-in" is used as no other option is remotely close
    *
+   * Under 4.0.0 the session is evaluated to classic.Session
+   *
+   * @param funcReg
+   * @param name
+   * @param builder
+   */
+  def registerFunction(sparkSession: SparkSession)(name: String, builder: Seq[Expression] => Expression) =
+    sparkSession.sessionState.functionRegistry.createOrReplaceTempFunction(name, builder, "built-in")
+
+  /**
+   * Registers functions with spark, Introduced in 0.4 - 3.2.0 support due to extra source parameter - "built-in" is used as no other option is remotely close
+   *
    * @param funcReg
    * @param name
    * @param builder
@@ -301,4 +313,34 @@ object ShimUtils {
     expressions.map(e => e.transformUp {
       case t: Stateful => t.freshCopy()
     })
+
+  /**
+   * Wrapper around call_function introduced in 3.5.0.  Spark 4 shims and above use internal.UnresolvedFunction
+   * and below uses analysis.UnresolvedFunction
+   * @param funcName
+   * @param cols
+   * @return
+   */
+  @scala.annotation.varargs
+  def callFunction(funcName: String, cols: Column*): Column =
+    column(com.sparkutils.shim.expressions.UnresolvedFunction4(funcName, cols.map(_.expr), false))
+
+  /**
+   * Returns true if the sparkSession is in classic mode, false for connect.  On pre 4.0.0 versions this
+   * always returns true.
+   * @param sparkSession
+   * @return
+   */
+  def isClassic(sparkSession: SparkSession): Boolean = true
+
+  /**
+   * Can this spark session be used.  On classic it's sparkSession.sqlContext.isStopped, on Spark 4 connect it's
+   * the private .isUsable function
+   *
+   * @param sparkSession
+   * @return
+   */
+  def isUsable(sparkSession: SparkSession): Boolean =
+    !sparkSession.sparkContext.isStopped
+
 }