+++ runtime/src/main/3.5.0-scala/com/sparkutils/shim/expressions/UnresolvedFunction.scala
@@ -22,7 +22,7 @@ object UnresolvedFunction5 {
     arguments: Seq[Expression],
     isDistinct: Boolean,
     filter: Option[Expression] = None,
-    ignoreNulls: Boolean = false) = UnresolvedFunction(nameParts.split("."), arguments, isDistinct, filter, ignoreNulls)
+    ignoreNulls: Boolean = false) = UnresolvedFunction(nameParts.split('.'), arguments, isDistinct, filter, ignoreNulls)
 
   def unapply(unresolvedFunction: UnresolvedFunction): Option[(String, Seq[Expression], Boolean, Option[Expression], Boolean)] =
     unresolvedFunction match {
@@ -42,13 +42,13 @@ object UnresolvedFunction4 {
    * @param isDistinct
    * @param filter
    * @param ignoreNulls
-   * @return
+   * @return 
    */
   def apply(
              nameParts: String,
              arguments: Seq[Expression],
              isDistinct: Boolean,
-             filter: Option[Expression] = None) = UnresolvedFunction(nameParts.split("."), arguments, isDistinct, filter)
+             filter: Option[Expression] = None) = UnresolvedFunction(nameParts.split('.'), arguments, isDistinct, filter)
 
   def unapply(unresolvedFunction: UnresolvedFunction): Option[(String, Seq[Expression], Boolean, Option[Expression])] =
     unresolvedFunction match {